Rename aesara.scan.opt to aesara.scan.rewriting
